Add test in main file
